requirements.txt
numpy==1.22.4
matplotlib==3.5.1
Pillow==9.1.1
torch==1.11.0
torchvision==0.12.0
tensorboard==2.5.0
tensorboardX==2.5.1
opencv-python==4.6.0.66
scipy==1.7.3
scikit-image==0.19.3

